server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
cryptography
httpx
pynacl
prometheus-client
pyyaml
yara-python
pandas
scikit-learn
pyod
faiss-cpu
chromadb
pytest
requests
llama-cpp-python
numpy
jsonschema
jinja2
scikit-learn
reportlab